auth_service/pyproject.toml
[project]
name = "auth-service"
version = "0.1.0"
description = "Authentication service with FastAPI, PostgreSQL, Redis, JWT and cookies"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",
    "redis>=5.0.0",
    "alembic>=1.13.0",
    "pydantic-settings>=2.4.0",
    "PyJWT>=2.9.0",
    "pwdlib[argon2]>=0.2.0",
    "psycopg[binary]>=3.3.3",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "ruff>=0.6.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
